SmartEducation/requirements.txt
appdirs==1.4.4
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
idna==3.10
markdown-it-py==3.0.0
materials==0.1.5
mdurl==0.1.2
numpy==2.3.2
packaging==25.0
phonenumberslite==9.0.11
pillow==11.3.0
pipdate==0.5.6
Pygments==2.19.2
requests==2.32.4
rich==14.1.0
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.5.0
